+++ Archspace2.Data/Resource.cs
@@ -6,9 +6,9 @@ namespace Archspace2
 {
     public class Resource
     {
-        public int ProductionPoint { get; set; }
-        public int ResearchPoint { get; set; }
-        public int MilitaryPoint { get; set; }
+        public long ProductionPoint { get; set; }
+        public long ResearchPoint { get; set; }
+        public long MilitaryPoint { get; set; }
 
         public Resource()
         {
